Major improvement to post-action stability checks - all methods have been extracted to individual module, OSGi bundles are verified after cq_package and cq_osgi_bundle, OSGi components after cq_osgi_config and cq_osgi_component
