+++ libraries/resource_cq_osgi_config.rb
@@ -51,7 +51,7 @@ class Chef
         @same_state_barrier = 3
         @error_state_barrier = 3
         @max_attempts = 30
-        @sleep_time = 2
+        @sleep_time = 3
       end
 
       def pid(arg = nil)
